Port ip-type-filter.tsx
