+++ src/proposals/comp/ip-type-filter.tsx
@@ -6,40 +6,28 @@
  */
 import React from "react"
 import * as _ from "lodash"
-import { Text, View } from "@nodegui/react-nodegui"
 import { observer } from "mobx-react-lite"
-import { CursorShape } from "@nodegui/nodegui"
+import styled from "styled-components"
 
 import { useStores } from "../../store"
 import { textSmall } from "../../ui-kit/typography"
 import { Toggle } from "../../ui-kit/toggle/toggle"
-import { brand } from "../../ui-kit/colors"
 
-const styleSheet = `
-#IpTypeFilter {
-    margin-top: 5;
-}
-#IpTypeFilter-Row-active {
-    width: 220;
-    height: 28;
-    background: "${brand}";
-    border-radius: 3px;
-}
-#IpTypeFilter-Row {
-    width: 220;
-    height: 28;
-    border-radius: 3px;
-}
-#IpTypeFilter-Row:hover {
-    background: #e6e6e6;
-}
-#IpTypeFilter-Row-Content {
-    padding-left: 10;
-    padding-right: 10;
-    width: 225;
-    flex-direction: "row";
-    justify-content: "space-between";
-}
+const Container = styled.div`
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+`
+
+const Title = styled.p`
+    ${textSmall}
+    color: #777;
+    margin: 12px;
+    margin-left: 12px;
+`
+
+const Count = styled.span`
+    margin-left: auto;
 `
 
 export const IpTypeFilter = observer(() => {
@@ -49,24 +37,8 @@ export const IpTypeFilter = observer(() => {
         return <></>
     }
     return (
-        <View
-            style={`
-            width: "100%";
-            padding: 2;
-            background: #fafafa;
-            border: 0;
-            flex-direction: column;
-            `}
-        >
-            <Text
-                style={`
-                ${textSmall}
-                color: #777;
-                margin: 5;
-            `}
-            >
-                IP type
-            </Text>
+        <Container>
+            <Title>IP type</Title>
             {Object.keys(ipTypeCounts)
                 .sort()
                 .map((ipType) => {
@@ -77,38 +49,12 @@ export const IpTypeFilter = observer(() => {
                     const count = ipTypeCounts[ipType]
                     const ipTypeDisplay = _.capitalize(ipType)
                     return (
-                        <View
-                            key={ipType}
-                            cursor={CursorShape.PointingHandCursor}
-                            id="IpTypeFilter"
-                            styleSheet={styleSheet}
-                        >
-                            <Toggle
-                                id={active ? "IpTypeFilter-Row-active" : "IpTypeFilter-Row"}
-                                onToggle={toggleAction}
-                            >
-                                <View id="IpTypeFilter-Row-Content">
-                                    <Text
-                                        style={`
-                                        color: ${active ? "white" : "inherit"};
-                                        `}
-                                    >
-                                        {ipTypeDisplay}
-                                    </Text>
-                                    <Text
-                                        style={`
-                                        height: 28;
-                                        qproperty-alignment: "AlignRight | AlignVCenter";
-                                        color: ${active ? "white" : "inherit"};
-                                        `}
-                                    >
-                                        {count}
-                                    </Text>
-                                </View>
-                            </Toggle>
-                        </View>
+                        <Toggle key={ipType} onClick={toggleAction} active={active}>
+                            {ipTypeDisplay}
+                            <Count>{count}</Count>
+                        </Toggle>
                     )
                 })}
-        </View>
+        </Container>
     )
 })
