+++ src/ui-kit/toggle/toggle.tsx
@@ -7,33 +7,18 @@
 import React from "react"
 import styled from "styled-components"
 
-export interface ToggleProps {
-    small?: boolean
-    children: React.ReactNode
-    active: boolean
-    onClick: Function
-}
+import { NavToggle, ToggleProps } from "./nav-toggle"
 
-export const Toggle = styled.div<ToggleProps>`
-    width: 152px;
-    height: ${(props: ToggleProps): string => (props.small ? "24px" : "28px")};
-    padding: 2px 12px;
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    user-select: none;
-
-    color: ${(props: ToggleProps): string => (props.active ? "#fff" : "#404040")};
+export const Toggle = styled(NavToggle)`
+    height: 28px;
+    width: calc(100% - 4px);
+    margin: 2px;
+    border-radius: 4px;
+    justify-content: flex-start;
     background: ${(props: ToggleProps): string =>
-        props.active ? "linear-gradient(180deg, #873a72 0%, #673a72 100%)" : "#fff"};
-    &:first-child {
-        border-top-left-radius: 4px;
-        border-bottom-left-radius: 4px;
-    }
-    &:last-child {
-        border-top-right-radius: 4px;
-        border-bottom-right-radius: 4px;
+        props.active ? "linear-gradient(180deg, #873a72 0%, #673a72 100%)" : "transparent"};
+    &:hover {
+        background: ${(props: ToggleProps): string =>
+            props.active ? "linear-gradient(180deg, #873a72 0%, #673a72 100%)" : "#e6e6e6"};
     }
 ` as React.FC<ToggleProps>
